+++ src/app/design-system/organisms/project-detail-card-skeleton/project-detail-card-skeleton.stories.ts
@@ -1,0 +1,11 @@
+import type { Meta, StoryObj } from '@storybook/angular';
+import { DsProjectDetailCardSkeleton } from './project-detail-card-skeleton';
+
+const meta: Meta<DsProjectDetailCardSkeleton> = {
+  title: 'Design System/Organisms/ProjectDetailCardSkeleton',
+  component: DsProjectDetailCardSkeleton,
+};
+export default meta;
+type Story = StoryObj<DsProjectDetailCardSkeleton>;
+
+export const Default: Story = {};